requirements.txt
httpx==0.27.2
backoff==2.2.1
pydantic==2.9.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
beautifulsoup4==4.12.3
html2text==2024.2.26
sqlite-utils==3.36
tqdm==4.66.5
jsonschema==4.23.0
pytest==8.3.3
